Reduce code duplication between broker-service and brokerd
