+++ lib/broker/CMakeLists.txt
@@ -36,6 +36,7 @@ else ()
         scavenger.hpp
         spawn.hpp
         uptime.hpp
+        utils.hpp
     )
 
     set (pstore_broker_lib_src
@@ -89,7 +90,7 @@ else ()
             INCLUDES "${pstore_broker_includes}"
         )
         target_link_libraries (pstore-broker PUBLIC
-            pstore-adt        
+            pstore-adt
             pstore-brokerface
             pstore-http
             pstore-json-lib
