+++ tools/brokerd/service/win32/broker_service.cpp
@@ -25,8 +25,6 @@
 
 #include "broker_service.hpp"
 
-#include <Dbghelp.h>
-
 #include "pstore/broker/command.hpp"
 #include "pstore/broker/gc.hpp"
 #include "pstore/broker/globals.hpp"
@@ -45,122 +43,12 @@
 #include "pstore/os/thread.hpp"
 #include "pstore/os/wsa_startup.hpp"
 #include "pstore/support/utf.hpp"
+#include "pstore/broker/utils.hpp"
 
 using namespace std::string_literals;
 using namespace pstore;
 
 extern romfs::romfs fs;
-
-namespace {
-
-    // create thread
-    // ~~~~~~~~~~~~~
-    template <class Function, class... Args>
-    auto create_thread (Function && f, Args &&... args)
-        -> std::future<typename std::result_of<Function (Args...)>::type> {
-
-        using return_type = typename std::result_of<Function (Args...)>::type;
-
-        std::packaged_task<return_type (Args...)> task (f);
-        auto fut = task.get_future ();
-
-        std::thread thr (std::move (task), std::forward<Args> (args)...);
-        thr.detach ();
-
-        return fut;
-    }
-
-    // thread init
-    // ~~~~~~~~~~~
-    void thread_init (std::string const & name) {
-        threads::set_name (name.c_str ());
-        create_log_stream ("broker." + name);
-    }
-
-    // create http worker thread
-    // ~~~~~~~~~~~~~~~~~~~~~~~~~
-    void create_http_worker_thread (gsl::not_null<std::vector<std::future<void>> *> const futures,
-                                    gsl::not_null<maybe<http::server_status> *> http_status,
-                                    bool announce_port) {
-        if (!http_status->has_value ()) {
-            return;
-        }
-        futures->push_back (create_thread (
-            [announce_port] (maybe<http::server_status> * const status) {
-                thread_init ("http");
-                PSTORE_ASSERT (status->has_value ());
-
-                http::channel_container channels{
-                    {"commits",
-                     http::channel_container_entry{&broker::commits_channel, &broker::commits_cv}},
-                    {"uptime",
-                     http::channel_container_entry{&broker::uptime_channel, &broker::uptime_cv}},
-                };
-
-                http::server (fs, &status->value (), channels,
-                              [announce_port] (in_port_t const port) {
-                                  if (announce_port) {
-                                      std::lock_guard<decltype (broker::iomut)> lock{broker::iomut};
-                                      std::cout << "HTTP listening on port " << port << std::endl;
-                                  }
-                              });
-            },
-            http_status.get ()));
-    }
-
-    // create worker threads
-    // ~~~~~~~~~~~~~~~~~~~~~
-    std::vector<std::future<void>> create_worker_threads (
-        std::shared_ptr<broker::command_processor> const & commands, brokerface::fifo_path & fifo,
-        std::shared_ptr<broker::scavenger> const & scav, std::atomic<bool> * const uptime_done) {
-
-        std::vector<std::future<void>> futures;
-
-        futures.push_back (create_thread ([&fifo, commands] () {
-            thread_init ("command");
-            commands->thread_entry (fifo);
-        }));
-
-        futures.push_back (create_thread ([scav] () {
-            thread_init ("scavenger");
-            scav->thread_entry ();
-        }));
-
-        futures.push_back (create_thread ([] () {
-            thread_init ("gcwatch");
-            broker::gc_process_watch_thread ();
-        }));
-
-        futures.push_back (create_thread (
-            [] (std::atomic<bool> * const done) {
-                thread_init ("uptime");
-                broker::uptime (done);
-            },
-            uptime_done));
-
-        return futures;
-    }
-
-    // get http server status
-    // ~~~~~~~~~~~~~~~~~~~~~~
-    /// Create an HTTP server_status object which reflects the user's choice of port.
-    decltype (auto) get_http_server_status (maybe<in_port_t> const & port) {
-        if (port.has_value ()) {
-            return maybe<http::server_status>{in_place, *port};
-        }
-        return nothing<http::server_status> ();
-    }
-
-    // make weak
-    // ~~~~~~~~~
-    /// Creates a weak_ptr from a shared_ptr. This can be done implicitly, but I want to make the
-    /// conversion explicit.
-    template <typename T>
-    std::weak_ptr<T> make_weak (std::shared_ptr<T> const & p) {
-        return {p};
-    }
-
-} // end anonymous namespace
 
 // (ctor)
 // ~~~~~~
@@ -170,9 +58,7 @@ broker_service::broker_service (TCHAR const * service_name, bool can_stop, bool 
 
 // (dtor)
 // ~~~~~~
-broker_service::~broker_service () {
-
-}
+broker_service::~broker_service () {}
 
 // start_handler
 // ~~~~~~~~~~~~~
@@ -187,7 +73,7 @@ void broker_service::start_handler (DWORD argc, TCHAR * argv[]) {
 
     std::pair<switches, int> opts = get_switches (argc, argv);
 
-    create_thread ([this, opts] { this->worker (opts.first); });
+    broker::create_thread ([this, opts] { this->worker (opts.first); });
 }
 
 
@@ -209,7 +95,7 @@ void broker_service::worker (switches opt) {
 
         std::vector<std::future<void>> futures;
         std::thread quit;
-        maybe<http::server_status> http_status = get_http_server_status (maybe<in_port_t>(8080));
+        maybe<http::server_status> http_status = broker::get_http_server_status (opt.http_port);
 
         std::atomic<bool> uptime_done{false};
 
@@ -226,10 +112,10 @@ void broker_service::worker (switches opt) {
                                    &http_status, &uptime_done);
 
         futures = create_worker_threads (commands, fifo, scav, &uptime_done);
-        create_http_worker_thread (&futures, &http_status, opt.num_read_threads);
+        broker::create_http_worker_thread (&futures, &http_status, opt.num_read_threads, fs);
 
         for (auto ctr = 0U; ctr < opt.num_read_threads; ++ctr) {
-            futures.push_back (create_thread ([ctr, &fifo, &record_file, commands] () {
+            futures.push_back (broker::create_thread ([ctr, &fifo, &record_file, commands] () {
                 auto const name = "read"s + std::to_string (ctr);
                 threads::set_name (name.c_str ());
                 create_log_stream ("broker." + name);
@@ -243,22 +129,22 @@ void broker_service::worker (switches opt) {
             f.get ();
         }
 
-        this->write_event_log_entry ("worker threads done, stopping quit thread", event_type::information);
+        this->write_event_log_entry ("worker threads done, stopping quit thread",
+                                     event_type::information);
         broker::notify_quit_thread ();
         quit.join ();
 
     } catch (std::exception const & ex) {
-        this->write_event_log_entry (std::string("error: ", ex.what()).c_str(), event_type::error);
-        this->set_service_status (SERVICE_STOPPED, EXIT_FAILURE);
-        return;
-    } catch (...) {
-        this->write_event_log_entry ("unknown error!",
+        this->write_event_log_entry (std::string ("error: ", ex.what ()).c_str (),
                                      event_type::error);
         this->set_service_status (SERVICE_STOPPED, EXIT_FAILURE);
         return;
+    } catch (...) {
+        this->write_event_log_entry ("unknown error!", event_type::error);
+        this->set_service_status (SERVICE_STOPPED, EXIT_FAILURE);
+        return;
     }
-    this->write_event_log_entry ("worker thread exiting",
-                                 event_type::information);
+    this->write_event_log_entry ("worker thread exiting", event_type::information);
     this->set_service_status (SERVICE_STOPPED, EXIT_SUCCESS);
 }
